refactor: move context and agent panel files to route directory
