+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-context-panel/context-panel-info.tsx
