+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-layout/dashboard-layout.tsx
@@ -9,9 +9,9 @@ import { useMatches } from "@tanstack/react-router";
 import { useStore } from "@tanstack/react-store";
 import type * as React from "react";
 import { useJobNotifications } from "@/features/notifications/use-job-notifications";
-import { ContextPanelTabContent } from "@/features/context-panel/context-panel";
-import { contextPanelStore } from "@/features/context-panel/context-panel-store";
-import { ContextPanelRail } from "@/features/context-panel/context-panel-rail";
+import { ContextPanelTabContent } from "../-context-panel/context-panel";
+import { ContextPanelRail } from "../-context-panel/context-panel-rail";
+import { contextPanelStore } from "../-context-panel/context-panel-store";
 import { AutoBugReporter } from "@/features/feedback/ui/auto-bug-reporter";
 import { MonthlySatisfactionSurvey } from "@/features/feedback/ui/monthly-satisfaction-survey";
 import { EarlyAccessProvider } from "@/hooks/use-early-access";
