+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-rubi-panel/agent-panel.tsx
