+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-rubi-panel/json-render/registry.tsx
