+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-rubi-panel/tool-call-card.tsx
