+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-rubi-panel/tool-renderers/advisor.tsx
