+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/-rubi-panel/tool-renderers/skill.tsx
