+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/services/$serviceId.tsx
@@ -12,7 +12,7 @@ import {
    closeContextPanel,
    openContextPanel,
    useContextPanelInfo,
-} from "@/features/context-panel/use-context-panel";
+} from "../../-context-panel/use-context-panel";
 import { orpc } from "@/integrations/orpc/client";
 import { ServiceBenefitsTab } from "./-services/service-benefits-tab";
 import { ServiceHeaderActions } from "./-services/service-header-actions";
