+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/services/benefits.tsx
@@ -55,7 +55,7 @@ import { DefaultHeader } from "@/components/default-header";
 import { requestTour } from "./-tour/store";
 import { TourHelpButton } from "./-tour/tour-help-button";
 import { QueryBoundary } from "@/components/query-boundary";
-import { useContextPanelInfo } from "@/features/context-panel/use-context-panel";
+import { useContextPanelInfo } from "../../-context-panel/use-context-panel";
 import { useOrgSlug, useTeamSlug } from "@/hooks/use-dashboard-slugs";
 import { useAlertDialog } from "@/hooks/use-alert-dialog";
 import { orpc } from "@/integrations/orpc/client";
